feat: rename package name #18
