--- app/src/main/java/com/example/mnrader/model/BarItem.kt
@@ -1,6 +1,0 @@
-package com.example.mnrader.model
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class BarItem (val title :String, val selectIcon: ImageVector, val onSelectedIcon :ImageVector, val route:String)
-
