--- app/src/main/java/com/example/mnrader/uicomponents/AddAnimals.kt
@@ -1,40 +1,0 @@
-package com.example.mnrader.uicomponents
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.AddCircle
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun AddAnimals(modifier: Modifier = Modifier) {
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        Icon(
-            imageVector = Icons.Default.AddCircle,
-            contentDescription = null,
-            tint = Color.Gray,
-            modifier = Modifier
-                .size(150.dp)
-                .align(Alignment.Center)
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun AddAnimalsPreview() {
-    MaterialTheme {
-        AddAnimals()
-    }
-}
