--- app/src/main/java/com/example/mnrader/uicomponents/Animals.kt
@@ -1,39 +1,0 @@
-package com.example.mnrader.uicomponents
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun Animals(modifier: Modifier = Modifier) {
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        Icon(
-            imageVector = Icons.Default.Person,
-            contentDescription = null,
-            tint = Color.Blue,
-            modifier = Modifier
-                .size(150.dp)
-                .align(Alignment.Center)
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun AnimalsPreview() {
-    MaterialTheme {
-        Animals()
-    }
-}
